l8/2 - private pages

+++ shows.php
@@ -15,7 +15,11 @@ $shows = $cmd->fetchAll();
 
 // start the list
 echo '<h1>Show Library</h1>';
-echo '<table><thead><th>Name</th><th>Release Year</th><th>Genre</th><th>Service</th><th>Actions</th></thead>';
+echo '<table><thead><th>Name</th><th>Release Year</th><th>Genre</th><th>Service</th>';
+if (!empty($_SESSION['username'])) {
+    echo '<th>Actions</th>';
+}
+echo '</thead>';
 
 // loop through the dataresult from the query, and display each show name
 foreach ($shows as $show) {
@@ -23,16 +27,18 @@ foreach ($shows as $show) {
         <td>' . $show['name'] . '</td>
         <td>' . $show['releaseYear'] . '</td>
         <td>' . $show['genre'] . '</td>
-        <td>' . $show['service'] . '</td>
-        <td class="actions">
-            <a href="edit-show.php?showId=' . $show['showId'] . '">
-                Edit
-            </a>&nbsp;
-            <a href="delete-show.php?showId=' .$show['showId'] . '" onclick="return confirmDelete();">
-                Delete
-            </a>
-        </td>
-        </tr>';
+        <td>' . $show['service'] . '</td>';
+        if (!empty($_SESSION['username'])) {
+            echo '<td class="actions">
+                <a href="edit-show.php?showId=' . $show['showId'] . '">
+                    Edit
+                </a>&nbsp;
+                <a href="delete-show.php?showId=' .$show['showId'] . '" onclick="return confirmDelete();">
+                    Delete
+                </a>
+            </td>';
+        }
+        echo '</tr>';
 }
 
 // end list
